AST: Fix `AvailabilityContext::forDeclSignature()`.
For declarations that aren't in source files there isn't an `AvailabilityScope`
tree to consult to find their `AvailabilityContext`. The fallback mechanism
used by `AvailabilityContext::forLocation()` walks up the `DeclContext`
hierarchy to build up the context for implicit code but this isn't entirely
correct since it skips over certain kinds of decls (like property
declarations). Instead, we need to walk up the parent decl hierarchy that is
used for availability inference to ensure that no decls with relevant
availability attributes are skipped.