Add presets for OpenBSD bootstrapping.
OpenBSD requires a minor quirk that differ from the standard
bootstrapping presets, namely, that because OpenBSD has a modified
LLVM toolchain by default, the build process should use that platform
toolchain and not the one just-built by the Swift process.

More specifically, Swift will still build against the Swift fork of
LLVM, but host binaries will be built with the platform toolchain.
This also necessitates using pure bridging instead of the default inline
bridging.

In my limited testing this has not proven to be problematic, and it will
be useful for future documentation updates to embed actual presets for
shared use.

Finally, we ensure that the destdir for these presets are /usr/local, not
/usr, as is conventional for software sourced outside of the platform;
these are not critical for bootstrapping but will ensure that less
changes will need to be made for creating installable packages. For now,
these are only added to stage1 and stage2; for some reason; adding this
to stage0 does not install properly.