[cxx-interop] Do not consider extensions blessed by SWIFT_PRIVATE_FILEID to be retroactive
Swift warns users when they conform an imported type to an imported
protocol, because such retroactive conformances are prone to producing
conflicting conformances in multiple modules.

When a user annotates a class as SWIFT_PRIVATE_FILEID, they are
explicitly conveying that extensions in the blessed file are part of
the class's private implementation (and thus privy to elevated access),
so we should also not treat conformances in that file as retroactive.

This patch fixes that and adds a test case.

rdar://148083096