Fix assertion in InstructionDeleter::trackIfDead.
This assertion was triggering when deleting scope-ending operations whose
operand had been replaced with an undef:

   end_borrow undef

Related to rdar://156431548

This was triggered during simplify-cfg by after adding blocks to the worklist,
then later cleaning up dead instructions in the worklist blocks. I'm unable to
create a test case based on simplify-cfg because the assertion is completely
unrelated to the optimization that adds blocks to the worklist.