[SILCombine] Bail load[copy](uteda) if move-only.
The pass rewrites `unchecked_take_enum_data_addr` whose uses are all
loads as loads--with the same ownership--of the base followed by
`unchecked_enum_data`s.  If the base is noncopyable and any of the loads
are `load [copy]`, performing this transformation would result in a
copy of that noncopyable base, which is invalid.  So bail if the base is
noncopyable and any of the loads are `load [copy]`s.

rdar://156543855