utils: Do not set the Androic C/C++ compilers
Due to a bug in CMake, the `project()` invocation for Android projects
always overrides the `CMAKE_C_COMPILER` and `CMAKE_CXX_COMPILER` passed
on the command line with the compilers from the NDK installation.

Since the values we passed on the command line were not taking effect,
these changes remove them from the cmake invocation. This fixes
incremental Android builds.