[Embedded] Introduce a new linkage model for Embedded Swift
This linkage model gives symbols generated in Embedded Swift
LLVM's linkonce_odr linkage (which allows the linker to merge
duplicates) unless they specifically need to be emitted as strong
symbols in the object. That fits better with creating and using static
libraries with Embedded Swift, and essentially maps to the idea of
"header-only" libraries in C and C++.

Introduce the attribute @alwaysEmitIntoObjectFile that forces code for
the symbol to be generated into the generated object file as a
non-mergeable symbol. Anything it depends on will also be emitted into
the object file, most of which will be via mergeable symbols based on
the rule above. Again, as a C/C++ analogy, this is like putting the
code into a ".c" or ".cpp" file, which will then include any
header-only code it depends on as well.

@alwaysEmitIntoObjectFile also suppresses SIL serialization for that
symbol, completely hiding the definition of the function from clients
of the module. This provides a rudimentary form of implementation
hiding that is common in "Desktop" Swift but was not available in
Embedded Swift.

All of this is behind an experimental feature flag,
EmbeddedLinkageModel, but is intended to subsume the existing
linkage model for Embedded Swift (which emits everying as a
non-mergeable symbol, so it doesn't work for static libraries), the
"mergeable symbols" experimental feature that's the existing model but
with "weak_odr" LLVM linkage, and the "-emit-empty-object-file" flag
(which emits nothing to the object file, and only works for libraries
but has no control).