Runtimes: ensure that we validate the linkage on Windows
Treat linker warnings as errors. This is to ensure that we do not
accidentally leak LNK4217 warnings which indicate miscompilation into
the runtimes. I thought that we had worked through them all, however,
one has seemed to have slipped through. The last remaining one is the
`_fatalErrorForwardModeDifferentiationDisabled` from the
`_Differentiation` module. This is incorrectly being marked as DLL
import:

```
LINK : warning LNK4217: symbol '_fatalErrorForwardModeDifferentiationDisabled' defined in 'DifferentiationUtilities-63c1ad.o' is imported by 'SIMDDifferentiation-e793ad.o' in function '$ss5SIMD2VyxSicisSBRzs10SIMDScalarRz16_Differentiation14DifferentiableRz13TangentVectorAdEPQzRszlTJfSUSpSr'
LINK : warning LNK4286: symbol '_fatalErrorForwardModeDifferentiationDisabled' defined in 'DifferentiationUtilities-63c1ad.o' is imported by 'FloatingPointDifferentiation-557cd1.o'
```