utils: add SBoM to Windows toolchain builds
This adds an additional bit of tooling to `build.ps1` to generate a SBoM
in SPDX (JSON) and cyclone (XML) formats. We also simultaneously
preserve the syft format for additional metadata. This is done in
preparation to ensure that we are able to track content beyond just the
manifest over time.