Introduce EmbeddedLeafLibrary feature for creating "leaf" modules
EmbeddedLeafLibrary infers @alwaysEmitInObjectFile for all public and
open entities in the module to which it is applied. Anything else will
retain linkonce_odr (mergeable) semantics. This allows the module to
"seal off" the Swift code such that the resulting object file can be
linked along with normal C code.

With this are a number of important bug fixes to the core of the new
Embedded Swift linkages model, including:
* Definitions from imported modules always get linkonce_odr linkage,
because they're likely to end up being duplicated by different clients
of the imported module file.
* Global variables also get linkonce_odr linkage unless they are
@alwaysEmitInObjectFile.

Now, building a library with the new embedded linkage model, by
default, produces effectively no symbols in the object file other than
linkonce_odr global variables. Symbols that are marked with
@alwaysEmitInObjectFile, or that have it inferred via
EmbeddedLeafLibrary, have strong definitions. Everything they depend
on is pulled in as a linkonce_odr symbol.