utils/wasm: Add Node.js WASI runner (#83524)
This runner can be utilized as an alternative to WasmKit when running tests or executables from Swift packages compiled to Wasm, when better performance or improved coverage of WASI 0.1 ABI is needed.

This script is meant to be running only locally at this time and doesn't assume presence of Node.js on CI, since no CI scripts were changed.