[concurrency] Add a new type Builtin.ImplicitActor.
This is currently not wired up to anything. I am going to wire it up in
subsequent commits.

The reason why we are introducing this new Builtin type is to represent that we
are going to start stealing bits from the protocol witness table pointer of the
Optional<any Actor> that this type is bitwise compatible with. The type will
ensure that this value is only used in places where we know that it will be
properly masked out giving us certainty that this value will not be used in any
manner without it first being bit cleared and transformed back to Optional<any
Actor>.

(cherry picked from commit 2fa3908e94c6519fd90a9ba634d6d78a25555d9b)

Conflicts:
	lib/AST/ASTPrinter.cpp
	lib/IRGen/IRGenModule.cpp