build.ps1: use pinned toolchain rather than built toolchain
Adjust the toolchain build to use the pinned toolchain rather than the
built toolchain. This is particularly important with the lack of ABI
stability on Windows. We now uniformly build the toolchain with the same
runtime. This extricates the SDK from the toolchain build.

Now that we build the toolchain properly against a SxS installable
version of the runtime, we can consume the runtime that we built
against and inject that into the toolchain image. This avoids the need
to rely on trying to maintain ABI compatibility in an ABI unstable
environment.