[cxx-interop] Export empty enums as namespaces
It is a common idiom in Swift to use empty enums as namespaces.
Previously, we just exported all zero sized types as unavailable
declarations so users had no way to access anything in these
types. This PR changes reverse interop to treat empty enums as
namespaces on the C++ side.

Note that we cannot treat enums as namespaces everywhere, an empty enum
nested in a class or struct would not work.

Moreover, this PR makes sure having empty enums as fields, parameters,
or return values does not regress.

It also covers some unlikely scenarios, like empty enums having an
initializer or an empty generic enum.

rdar://147499583