ComputeSideEffects: compute properties even if function has an effect attribute.
For the bug this patch aims to fix, we only need to compute the isDeinitBarrier
property. We could attempt to only store the computed value of the
isDeinitBarrier property, but there is no benefit to this over storing all the
computed values. The benefit of the previous behaviour was avoiding the overhead
of computing any properties in the first place.

rdar://155870190

This aims to resolve the same problem as #38324.