nfc: refactor Backtracing/Crash.swift
This test fused both the CHECK directives and the
source code into one file, and is sensitive to the line numbers
that source code appears on in that file. Thus, any attempts to
add new lines to this test would break it.

This refactoring moves the source code to a separate file so
its line numbers stay stable.